// src/lib/orders/model/internalStatus.ts
export enum InternalStatus {
	ImBestellkorb = 0,
	NeuEingegangen = 1,
	PruefungEinsichtsgesuch = 2,
	EinsichtsgesuchWeitergeleitet = 3,
	PruefungFreigabe = 4,
	Abgebrochen = 5,
	FuerDigitalisierungBereit = 6,
	FuerAushebungBereit = 7,
	AushebungsauftragErstellt = 8,
	Ausgeliehen = 9,
	DigitalisierungAbgebrochen = 10,
	DigitalisierungExtern = 11,
	ZumReponierenBereit = 12,
	Abgeschlossen = 13,
}